Fix some things that broke in recent commits

+++ src/boe.locutils.h
@@ -22,7 +22,7 @@ bool is_container(location loc);
 void update_explored(location dest);
 bool is_blocked(location to_check);
 bool monst_on_space(location loc,short m_num);
-short monst_there(location where) ;
+size_t monst_there(location where) ;
 bool monst_can_be_there(location loc,short m_num);
 bool monst_adjacent(location loc,short m_num);
 bool monst_can_see(short m_num,location l);
